+++ spec/unit/publish_artifacts_spec.rb
@@ -42,17 +42,19 @@ describe KubeDeployTools::PublishArtifacts do
 
     # images.yaml, tarballs, and bare deploy.yml to test extra file
     # support
-    expected_uploads = Set.new([
+    expected_uploads = [
       'manifests_colo-service_prod_default.tar.gz',
       'manifests_colo-service_staging_default.tar.gz',
-      'manifests_gcp_prod_default.tar.gz',
       'manifests_local_staging_default.tar.gz',
       'manifests_us-east-1_prod_default.tar.gz',
       'manifests_us-east-1_staging_default.tar.gz',
+      'manifests_ingestion_prod_default.tar.gz',
+      'manifests_pippio-production_prod_default.tar.gz',
+      'manifests_platforms_prod_default.tar.gz',
       'deploy.yml',
       'images.yaml',
-    ])
-    expect(uploads).to eq(expected_uploads)
+    ]
+    expect(uploads).to contain_exactly(*expected_uploads)
   end
 end